pyproject.toml
[project]
name = "glyphatlas-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paddleocr==3.3.3",
    "paddlepaddle==3.0.0",
    "pillow>=12.1.0",
    "protobuf==6.33.4",
    "pymupdf>=1.26.7",
]
